fix: change path stat page
