+++ src/components/NewsDrawer/Events/expansionItems/expansion-item-stats.vue
