+++ src/components/NewsDrawer/NewsDrawer.vue
@@ -3,7 +3,6 @@
     <news-drawer-home />
     <news-drawer-club />
     <news-drawer-event />
-    <news-drawer-stats />
     <news-drawer-archive />
     <news-drawer-fanat />
     <news-drawer-contact />
@@ -19,11 +18,10 @@ import NewsDrawerFanat from 'components/NewsDrawer/Fanats/newsDrawerFanat.vue'
 import NewsDrawerContact from 'components/NewsDrawer/Contacts/newsDrawerContact.vue'
 import NewsDrawerSite from 'components/NewsDrawer/Site/newsDrawerSite.vue'
 import NewsDrawerArchive from 'components/NewsDrawer/Archives/newsDrawerArchive.vue'
-import NewsDrawerStats from 'components/NewsDrawer/Stats/news-drawer-stats.vue'
 
 export default {
   name: 'NewsDrawer',
-  components: { NewsDrawerStats, NewsDrawerArchive, NewsDrawerSite, NewsDrawerContact, NewsDrawerFanat, NewsDrawerHome, NewsDrawerEvent, NewsDrawerClub }
+  components: { NewsDrawerArchive, NewsDrawerSite, NewsDrawerContact, NewsDrawerFanat, NewsDrawerHome, NewsDrawerEvent, NewsDrawerClub }
 }
 </script>
